backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
pyodbc==5.1.0
python-dotenv==1.0.1
